Use correct glsl type casting
git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@580 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/shader/vv_intersection.glsl
@@ -18,8 +18,8 @@ void main()
   for (int i=0; i<4; ++i)
   {
     // int(brickDimInv[3] == frontIndex.
-    int vIdx1 = sequence[(int)(brickDimInv[3] * 8 + v1[gl_Vertex.x*4+i])];
-    int vIdx2 = sequence[(int)(brickDimInv[3] * 8 + v2[gl_Vertex.x*4+i])];
+    int vIdx1 = sequence[int(brickDimInv[3] * 8.0 + float(v1[int(gl_Vertex.x)*4+i]))];
+    int vIdx2 = sequence[int(brickDimInv[3] * 8.0 + float(v2[int(gl_Vertex.x)*4+i]))];
 
     vec3 vecV1 = vertices[vIdx1];
     vec3 vecV2 = vertices[vIdx2];
